changes as suggested by reviewer
-class BlockStorgae has its own file
-TreeIterator::levels reverted to std::stack (get rid of vectorOfRanges)
-ContourPointsStorage and ContourCodesStorage are template specialization of the same class ContourDataStorage
-ContourPointsStorage and ContourCodesStorage us a static size of 0 for their inner BlockStorage

I have run the tstandard findContours perf_tests to check that all those changes were not a slowdown for the average case (because the current use case of the PR must not take precedence). It's ok.
