+++ modules/imgproc/src/contours_common.hpp
@@ -7,7 +7,7 @@
 
 #include "precomp.hpp"
 
-#include <array>
+#include "contours_blockstorage.hpp"
 
 namespace cv {
 
@@ -27,156 +27,6 @@ inline schar clamp_direction(schar dir)
 {
     return std::min(dir, (schar)15);
 }
-
-// BLOCK_SIZE_ELEM - number of elements in a block
-// STATIC_CAPACITY_BYTES - static memory in bytes for preallocated blocks
-template <typename T, size_t BLOCK_SIZE_ELEM = 1024, size_t STATIC_CAPACITY_BYTES = 4096>
-class BlockStorage {
-    public:
-        using value_type = T;
-        typedef struct {value_type data[BLOCK_SIZE_ELEM];} block_type;
-
-        BlockStorage()
-        {
-            const size_t minDynamicBlocks = !staticBlocksCount ? 1 : 0;
-            for(size_t i = 0 ; i<minDynamicBlocks ; ++i)
-                dynamicBlocks.push_back(new block_type);
-        }
-        BlockStorage(const BlockStorage&) = delete;
-        BlockStorage(BlockStorage&&) noexcept = default;
-        ~BlockStorage() {
-            for(const auto & block : dynamicBlocks) {
-                delete block;
-            }
-        }
-        BlockStorage& operator=(const BlockStorage&) = delete;
-        BlockStorage& operator=(BlockStorage&&) noexcept = default;
-
-        void clear(void) {
-            const size_t minDynamicBlocks = !staticBlocksCount ? 1 : 0;
-            for(size_t i = minDynamicBlocks, count = dynamicBlocks.size() ; i<count ; ++i ) {
-                delete dynamicBlocks[i];
-            }
-            dynamicBlocks.resize(minDynamicBlocks);
-            sz = 0;
-        }
-
-        void push_back(const value_type& value) {
-            const size_t blockIndex = sz / BLOCK_SIZE_ELEM;
-            const size_t currentBlocksCount = staticBlocksCount+dynamicBlocks.size();
-            if (blockIndex == currentBlocksCount)
-                dynamicBlocks.push_back(new block_type);
-            block_type& cur_block =
-                (blockIndex < staticBlocksCount) ? staticBlocks[blockIndex] :
-                *dynamicBlocks[blockIndex-staticBlocksCount];
-            cur_block.data[sz % BLOCK_SIZE_ELEM] = value;
-            ++sz;
-        }
-
-        size_t size() const { return sz; }
-
-        const value_type& at(size_t index) const {
-            const size_t blockIndex = index / BLOCK_SIZE_ELEM;
-            const block_type& cur_block =
-                (blockIndex < staticBlocksCount) ? staticBlocks[blockIndex] :
-                *dynamicBlocks[blockIndex-staticBlocksCount];
-            return cur_block.data[index % BLOCK_SIZE_ELEM];
-        }
-        value_type& at(size_t index) {
-            const size_t blockIndex = index / BLOCK_SIZE_ELEM;
-            block_type& cur_block =
-                (blockIndex < staticBlocksCount) ? staticBlocks[blockIndex] :
-                *dynamicBlocks[blockIndex-staticBlocksCount];
-            return cur_block.data[index % BLOCK_SIZE_ELEM];
-        }
-        const value_type& operator[](size_t index) const {return at(index);}
-        value_type& operator[](size_t index) {return at(index);}
-    public:
-        friend class RangeIterator;
-        class RangeIterator
-        {
-            public:
-                RangeIterator(const BlockStorage* _owner, size_t _first, size_t _last)
-                             :owner(_owner),remaining(_last-_first),
-                              blockIndex(_first/BLOCK_SIZE_ELEM),offset(_first%BLOCK_SIZE_ELEM) {
-                }
-            private:
-                const BlockStorage* owner = nullptr;
-                size_t remaining = 0;
-                size_t blockIndex = 0;
-                size_t offset = 0;
-            public:
-                bool done(void) const {return !remaining;}
-                std::pair<const value_type*, size_t> operator*(void) const {return get();}
-                std::pair<const value_type*, size_t> get(void) const {
-                    const block_type& cur_block =
-                        (blockIndex < owner->staticBlocksCount) ? owner->staticBlocks[blockIndex] :
-                        *owner->dynamicBlocks[blockIndex-owner->staticBlocksCount];
-                    const value_type* rangeStart = cur_block.data+offset;
-                    const size_t rangeLength = std::min(remaining, BLOCK_SIZE_ELEM-offset);
-                    return std::make_pair(rangeStart, rangeLength);
-                }
-                RangeIterator& operator++() {
-                    std::pair<const value_type*, size_t> range = get();
-                    remaining -= range.second;
-                    offset = 0;
-                    ++blockIndex;
-                    return *this;
-                }
-        };
-        RangeIterator getRangeIterator(size_t first, size_t last) const {
-          return RangeIterator(this, first, last);
-        }
-    private:
-        std::array<block_type, STATIC_CAPACITY_BYTES/(BLOCK_SIZE_ELEM*sizeof(value_type))> staticBlocks;
-        const size_t staticBlocksCount = STATIC_CAPACITY_BYTES/(BLOCK_SIZE_ELEM*sizeof(value_type));
-        std::vector<block_type*> dynamicBlocks;
-        size_t sz = 0;
-};
-
-template<typename T>
-class vectorOfRanges
-{
-    public:
-        vectorOfRanges(void) = default;
-        vectorOfRanges(const vectorOfRanges&) = default;
-        vectorOfRanges(vectorOfRanges&& other) noexcept = default;
-        ~vectorOfRanges() = default;
-    public:
-        vectorOfRanges& operator=(const vectorOfRanges&) = default;
-        vectorOfRanges& operator=(vectorOfRanges&& other) noexcept = default;
-    public:
-        bool empty(void) const {return !_size;}
-        size_t size(void) const {return _size;}
-        T at(size_t index) const {
-            for(const auto& range : _ranges) {
-                if (index < range.second)
-                    return static_cast<T>(range.first+index);
-                else
-                    index -= range.second;
-          }
-          return _ranges[0].first;//should not occur
-        }
-        T back(void) const {return at(_size-1);}
-    public:
-        void push_back(const T& value) {
-            if (_ranges.empty() || (value != back()+1))
-                _ranges.push_back(std::make_pair(value, 1));
-            else
-                ++_ranges.back().second;
-            ++_size;
-        }
-        void pop_back(void) {
-            if (_ranges.back().second == 1)
-                _ranges.pop_back();
-            else
-                --_ranges.back().second;
-            --_size;
-        }
-    private:
-        std::vector<std::pair<T, size_t> > _ranges;
-        size_t _size = 0;
-};
 
 template <typename T>
 class TreeNode
@@ -297,10 +147,10 @@ template <typename T>
 class TreeIterator
 {
 public:
-    TreeIterator(Tree<T>& tree_) : tree(tree_)//,levels(&tree._treeIteratorArena)
+    TreeIterator(Tree<T>& tree_) : tree(tree_)
     {
         CV_Assert(!tree.isEmpty());
-        levels.push_back(0);
+        levels.push(0);
     }
     bool isDone() const
     {
@@ -308,13 +158,13 @@ public:
     }
     const TreeNode<T>& getNext_s()
     {
-        int idx = levels.back();
-        levels.pop_back();
+        int idx = levels.top();
+        levels.pop();
         const TreeNode<T>& res = tree.elem(idx);
         int cur = tree.lastSibling(res.first_child);
         while (cur != -1)
         {
-            levels.push_back(cur);
+            levels.push(cur);
             cur = tree.elem(cur).prev;
         }
         return res;
@@ -322,96 +172,58 @@ public:
 
 private:
     Tree<T>& tree;
-    vectorOfRanges<int> levels;
+    std::stack<int> levels;
 };
 
 //==============================================================================
 
-class ContourPointsStorage
+template <typename T, size_t BLOCK_SIZE_ELEM, size_t STATIC_CAPACITY_BYTES>
+class ContourDataStorage
 {
-    public:
-        typedef cv::Point point_storage_t;
-        typedef BlockStorage<point_storage_t> storage_t;
-    public:
-        ContourPointsStorage(void) = delete;
-        ContourPointsStorage(storage_t* _storage):storage(_storage) {}
-        ContourPointsStorage(const ContourPointsStorage&) = delete;
-        ContourPointsStorage(ContourPointsStorage&&) noexcept = default;
-        ~ContourPointsStorage() = default;
-        ContourPointsStorage& operator=(const ContourPointsStorage&) = delete;
-        ContourPointsStorage& operator=(ContourPointsStorage&&) noexcept = default;
-    public:
-        storage_t::RangeIterator getRangeIterator(void) const {return storage->getRangeIterator(first, last);}
-    public:
-        bool empty(void) const {return first == last;}
-        size_t size(void) const {return last - first;}
-    public:
-        void clear(void) {first = last;}
-        bool resize(size_t newSize) {
-            bool ok = (newSize <= size());
-            if (ok)
-                last = first+newSize;
-            return ok;
+public:
+    typedef T data_storage_t;
+    typedef BlockStorage<data_storage_t, BLOCK_SIZE_ELEM, STATIC_CAPACITY_BYTES> storage_t;
+public:
+    ContourDataStorage(void) = delete;
+    ContourDataStorage(storage_t* _storage):storage(_storage) {}
+    ContourDataStorage(const ContourDataStorage&) = delete;
+    ContourDataStorage(ContourDataStorage&&) noexcept = default;
+    ~ContourDataStorage() = default;
+    ContourDataStorage& operator=(const ContourDataStorage&) = delete;
+    ContourDataStorage& operator=(ContourDataStorage&&) noexcept = default;
+public:
+    typename storage_t::RangeIterator getRangeIterator(void) const {return storage->getRangeIterator(first, last);}
+public:
+    bool empty(void) const {return first == last;}
+    size_t size(void) const {return last - first;}
+public:
+    void clear(void) {first = last;}
+    bool resize(size_t newSize) {
+        bool ok = (newSize <= size());
+        if (ok)
+            last = first+newSize;
+        return ok;
+    }
+    void push_back(const data_storage_t& value) {
+        if (empty()) {
+            first = storage->size();
         }
-        void push_back(const point_storage_t& value) {
-            if (empty()) {
-                first = storage->size();
-            }
-            storage->push_back(value);
-            last = storage->size();
-        }
-        const cv::Point& at(size_t index) const {return storage->at(first+index);}
-        cv::Point& at(size_t index) {return storage->at(first+index);}
-        const cv::Point& operator[](size_t index) const {return at(index);}
-        cv::Point& operator[](size_t index) {return at(index);}
-    private:
-        storage_t* storage = nullptr;
-        size_t first = 0;
-        size_t last = 0;
+        storage->push_back(value);
+        last = storage->size();
+    }
+    const data_storage_t& at(size_t index) const {return storage->at(first+index);}
+    data_storage_t& at(size_t index) {return storage->at(first+index);}
+    const data_storage_t& operator[](size_t index) const {return at(index);}
+    data_storage_t& operator[](size_t index) {return at(index);}
+private:
+    storage_t* storage = nullptr;
+    size_t first = 0;
+    size_t last = 0;
 };
 
-class ContourCodesStorage
-{
-    public:
-        typedef schar code_storage_t;
-        typedef BlockStorage<code_storage_t, 1024, 0> storage_t;
-    public:
-        ContourCodesStorage(void) = delete;
-        ContourCodesStorage(storage_t* _storage):storage(_storage) {}
-        ContourCodesStorage(const ContourCodesStorage&) = delete;
-        ContourCodesStorage(ContourCodesStorage&&) noexcept = default;
-        ~ContourCodesStorage() = default;
-        ContourCodesStorage& operator=(const ContourCodesStorage&) = delete;
-        ContourCodesStorage& operator=(ContourCodesStorage&&) noexcept = default;
-    public:
-        storage_t::RangeIterator getRangeIterator(void) const {return storage->getRangeIterator(first, last);}
-    public:
-        bool empty(void) const {return first == last;}
-        size_t size(void) const {return last - first;}
-    public:
-        void clear(void) {first = last;}
-        bool resize(size_t newSize) {
-            bool ok = (newSize <= size());
-            if (ok)
-                last = first+newSize;
-            return ok;
-        }
-        void push_back(const code_storage_t& value) {
-            if (empty()) {
-                first = storage->size();
-            }
-            storage->push_back(value);
-            last = storage->size();
-        }
-        const schar& at(size_t index) const {return storage->at(first+index);}
-        schar& at(size_t index) {return storage->at(first+index);}
-        const schar& operator[](size_t index) const {return at(index);}
-        schar& operator[](size_t index) {return at(index);}
-    private:
-        storage_t* storage = nullptr;
-        size_t first = 0;
-        size_t last = 0;
-};
+typedef ContourDataStorage<cv::Point, 1024, 0> ContourPointsStorage;
+typedef ContourDataStorage<schar, 1024, 0> ContourCodesStorage;
+
 class Contour
 {
 public:
